+alts adjustment
Adjusted alts to allow for some security in setting alts by adding a confirmation command. This will @mail someone a confirmation code that they have to use to confirm they are the alt of someone else. Additionally, I've added the ability to block someone from adding you as an alt as a just-in-case.