Fix capitalization
There were some issues about capitalization, especially with stats that are an acronym (PTSD, OCPD, OCD). This addresses that. Additionally, we had some other issues with capitalization following certain punctuation. This makes sure that stats like Power-Brokering and Primal-Urge retain the capitalization following the dash.

When calling selfstat, the command will now check for anything that might have an incorrect capitalization.

This all applies to staffstat as well.

Additionally, there is now a +fixcaps command that fixes the caps issues on PCs. The syntax is +fixcaps <name>