Shifter-Specific Merits
There was an error in shifter-specific merits and flaws that threw a typeclass error due to two reasons:
1) When no shifter type is set, it didn't default back to the original char_type (splat)
2) When a shifter type was set, we were evaluating the restriction as a string ('shifter_type') rather than pulling from the allowed_types dictionary or using the char_type function.

Notably 'shifter_type' isn't the actual name of the stat; it's just 'type.' I didn't update that since adding in the char_type to both sections of selfstat and allowed_types to the secondary section fixed the issue for multiple changing breed types.

As a result of this fix, shifter types are now evaluated when attempting to set merits and flaws (go figure).