Update load_wod20th_stats.py
Added option in load_wod20th_stats to indicate a delay between files when running the command from bash. This works as follows:

```evennia load_wod20th_stats --delay [time_in_seconds]```

This will run all files in the data folder with the stated delay. The default delay is 5 seconds. The intention for this is to prevent the database from locking up due to too many queries when running all files on the system. This is particularly an issue with sqlite3.